llm-quiz-project/requirements.txt
flask==3.0.0
openai==1.54.0
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.0
python-dotenv==1.0.0
gunicorn==21.2.0
openpyxl==3.1.2
lxml==5.1.0